Fixed bug: int32 is long not int!

+++ MCAL_layer/mcal_std_types.h
@@ -34,9 +34,9 @@
 /*************** data types ****************/
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned int uint32;
+typedef unsigned long uint32;
 typedef signed char sint8;
-typedef signed int sint32;
+typedef signed long sint32;
 typedef signed short sint16;
 typedef uint8 STD_ReturnType;
 /*************** function declarations *****/
